move store inside lib folder
