+++ frontend/components/Navbar.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { useAuthStore } from "@/store/authStore"
+import { useAuthStore } from "@/lib/store/authStore"
 import { useRouter, usePathname } from "next/navigation"
 import { useEffect } from "react"
 import ThemeToggle from "@/components/ThemeToggle"
